Yikes!  I reversed all the pins on the gps connector.
I didn't realize the plug had to be turned
upside down to fit it into the connector.

The kicad files have been updated.  It also seems
I've swapped the transmit and receive on the gps
connector.  I've updated the code to compensate.
Found compatibility problems when using a windows
version of make.  Updated the makefile.